Switches from 16 bit representation to 32 bit

+++ Mutator.h
@@ -14,7 +14,7 @@ public:
     static void mutateProgram(Program& program);
     static std::pair<Program, Program> crossover(const Program &program1, const Program &program2);
 private:
-    static void mutateInstruction(uint16_t& instruction);
+    static void mutateInstruction(uint32_t& instruction);
     static std::mt19937 gen;
     static std::uniform_real_distribution<> rng;
 };
